requirement.txt
ipykernel
fastapi
numpy
black
python=3.10.8
beautifulsoup4
tqdm
requests
pip
openai-whisper
langchain
weaviate-client
sqlalchemy
isort
transformers
gradio
python-dotenv
llama-index
youtube-transcript-api
